# doc/api.yaml
openapi: 3.0.0
paths:
  /user:
    get:
      operationId: UserController_findAll
      summary: Get all users
      description: Get all users
      parameters: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      tags: &ref_0
        - User
    post:
      operationId: UserController_create
      summary: Create user
      description: Create a new user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Body does not contain required fields
      tags: *ref_0
  /user/{userId}:
    get:
      operationId: UserController_findOne
      summary: Get single user by id
      description: Get single user by id
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User id is invalid (not uuid)
        '404':
          description: User not found
      tags: *ref_0
    put:
      operationId: UserController_update
      summary: Update a user's password
      description: Updates a user's password by ID
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordDto'
      responses:
        '200':
          description: The user has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: User id is invalid (not uuid)
        '403':
          description: Old password is wrong
        '404':
          description: User not found
      tags: *ref_0
    delete:
      operationId: UserController_remove
      summary: Delete user
      description: Deletes user by ID
      parameters:
        - name: userId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: The user has been deleted
        '400':
          description: User id is invalid (not uuid)
        '404':
          description: User not found
      tags: *ref_0
...
info:
  title: Home Library Service
  description: Home music library service
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        login:
          type: string
          example: TestUser
        version:
          type: number
          example: 1
        createdAt:
          type: number
          example: 1655000000
        updatedAt:
          type: number
          example: 1655000000
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: string
          description: The user's login
        password:
          type: string
          description: The user's password
      required:
        - login
        - password
    UpdatePasswordDto:
      type: object
      properties:
        oldPassword:
          type: string
          description: The user's old password
        newPassword:
          type: string
          description: The user's new password
      required:
        - oldPassword
        - newPassword
...